+++ geotrek/api/v2/views/__init__.py
@@ -20,7 +20,8 @@ if 'geotrek.sensitivity' in settings.INSTALLED_APPS:
     from .sensitivity import SportPracticeViewSet  # noqa
     from .sensitivity import SpeciesViewSet  # noqa
 if 'geotrek.tourism' in settings.INSTALLED_APPS:
-    from .tourism import TouristicContentViewSet, TouristicEventViewSet, TouristicEventTypeViewSet, InformationDeskViewSet, InformationDeskTypeViewSet, TouristicContentCategoryViewSet, LabelAccessibilityViewSet, TouristicEventPlaceViewSet  # noqa
+    from .tourism import (TouristicContentViewSet, TouristicEventViewSet, TouristicEventTypeViewSet, InformationDeskViewSet, InformationDeskTypeViewSet,  # noqa
+                         TouristicContentCategoryViewSet, LabelAccessibilityViewSet, TouristicEventPlaceViewSet, TouristicEventOrganizerViewSet)  # noqa
 if 'geotrek.zoning' in settings.INSTALLED_APPS:
     from .zoning import CityViewSet, DistrictViewSet  # noqa
 if 'geotrek.outdoor' in settings.INSTALLED_APPS: