+++ docs/changelog.rst
@@ -20,6 +20,7 @@ CHANGELOG
 - Add a multiselect to filter the Blades by more than one manager
 - Filter by begin date by default on touristic events in APIv2 (#3597)
 - Add model LinePictogram for each line (#3327)
+- Create Organizer model for touristic events, configurable in admin site (#3625)
 
 **Documentation**